requirements.txt
alembic==1.10.4
Flask==2.3.2
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
numpy==1.24.3
opencv_python_headless==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
rembg==2.0.36
SQLAlchemy==2.0.13
Werkzeug==2.3.4
WTForms==3.0.1
